version1

+++ Backend/Crowdfunding/src/main/java/com/example/Crowdfunding/Admins/Admin.java
@@ -13,7 +13,7 @@ public class Admin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "SEQUENCE_ADMINS")
     @SequenceGenerator(name = "SEQUENCE_ADMINS", sequenceName = "admins_admin_id_seq", allocationSize = 1)
-    private Long admin_id;
+    private int admin_id;
 
     private String email;
     private String password;
